Retry on inherited exceptions as well
The existing retry mechanism almost never triggered when the underlying
function call throws an exception. In particular, it never retries to
call a function if it fails with a connection error.

The mechanics was broken by `7249ad7 Add BulkProcessor` commit. So,
there should not be any undesirable side-effects from retrying on all
exceptions by default.